==> code/12b.nitrogen-shoreline-spatial-modeling.qmd <==
---
title: "Spatial regression - ammonia, nitrate + nitrite, phosphate"
format: 
  html:
    self-contained: true
editor: source
editor_options: 
  chunk_output_type: console
---

## Data prep

```{r data, warning=F, message=F}
#| code-fold: true

library(tidyverse)
library(phyloseq)
library(spmodel)
library(sf)
library(emmeans)
library(ggbeeswarm)
library(ggsignif)

shore_colors <- c('#1f78b4', '#fb9a99', '#33a02c')

level <- 'family'; level2 <- 'Family'

nutrients <- 
  readr::read_rds(paste0('./data/ati-physeq-rrfy-', level2, '.rds')) %>% 
  sample_data() %>% 
  data.frame() %>% 
  dplyr::select(Year, Site, Habitat, Island_shore, Ammonia,  
                Nitrite_plus_Nitrate, Phosphate, Latitude, Longitude) %>% 
  dplyr::mutate(Year = as.factor(Year)) %>% 
  na.omit() %>% 
  st_as_sf(coords = c("Longitude", "Latitude"), crs = 4326) %>% 
  st_transform(crs = 3305)

boxplot_theme <-
  theme(
    axis.title.x = element_text(size = 16, face = "bold"),
    axis.title.y = element_text(size = 16, face = "bold"),
    axis.text.x  = element_text(size = 14),
    axis.text.y  = element_text(size = 14),
    legend.position = 'none'
  )

ggplot(data = nutrients) + 
  geom_boxplot(aes(x = Island_shore,
                   y = Ammonia)) +
  facet_wrap(~ Year) +
  theme_bw()


```

## Model and plot ammonia

```{r ammonia, warning=F, message=F, eval=T}
#| code-fold: true

formula <-
  Ammonia ~
  Year + 
  Island_shore

sp.mod1 <- splm(formula = formula, 
               data = nutrients,
               random = ~ Site,
               estmethod = 'ml',
               spcov_type = 'gaussian') 

formula <-
  Ammonia ~
  Year * 
  Island_shore

sp.mod <- splm(formula = formula, 
               data = nutrients,
               random = ~ Site,
               estmethod = 'ml',
               spcov_type = 'gaussian') 

summary(sp.mod)

glances(sp.mod1, sp.mod2)

em.ammonia <- emmeans::emmeans(sp.mod, ~ Year)

print(em.ammonia)

pairs(em.ammonia)

plot(pairs(em.ammonia)) +
  theme_bw() +
  geom_vline(xintercept = 0)

em.ammonia <- emmeans::emmeans(sp.mod, ~ Island_shore, by = 'Year')

print(em.ammonia)

pairs(em.ammonia)

plot(pairs(em.ammonia)) +
  theme_bw() +
  geom_vline(xintercept = 0)

nutrients$Year <- factor(nutrients$Year, levels = c("2021","2022","2023"))

## 1) Tidy pairwise tests for ggsignif
pairs_df <- 
  pairs(em.ammonia) %>%
  summary(adjust = "tukey") %>%
  as_tibble() %>%
  separate(contrast, into = c("group1", "group2"), sep = " - ") %>%
  mutate(
    # make sure group names match x-axis categories exactly
    group1 = as.character(group1),
    group2 = as.character(group2),
    label  = paste0("p=", scales::pvalue(p.value))) %>% 
  mutate(label = ifelse(label == 'p=<0.001', 'p<0.001', label))

## 2) Give each comparison a y position per Year so brackets don’t overlap
yr_tops <- nutrients %>%
  group_by(Year) %>%
  summarise(y_top = max(Ammonia, na.rm = TRUE), .groups = "drop")

range_y <- diff(range(nutrients$Ammonia, na.rm = TRUE))

pairs_df <- pairs_df %>%
  left_join(yr_tops, by = "Year") %>%
  group_by(Year) %>%
  mutate(
    # stack 3 comparisons above the tallest box for that Year
    y_position = y_top + (row_number()) * (0.06 * range_y)
  ) %>%
  ungroup()

## 3) Plot with ggsignif (manual = TRUE)
p <-
  ggplot(nutrients, aes(x = Island_shore, y = Ammonia, color = Island_shore)) +
  geom_boxplot(outlier.shape = NA, alpha = 0.9, size=0.8) +
  geom_quasirandom(size = 2.5, alpha = 0.3, stroke = NA) +
  scale_color_manual(values = shore_colors) +
  facet_wrap(~ Year) +
  ggsignif::geom_signif(
    data = pairs_df,
    aes(xmin = group1, xmax = group2, annotations = label, y_position = y_position),
    inherit.aes = FALSE,
    manual = TRUE,          # <-- REQUIRED when mapping via data frame
    tip_length = 0.01,
    vjust = 0.2,
    textsize = 3.6) +
  labs(y = expression("Ammonia ("*mu*"M)"),
       x = 'Island Shore') +
  theme_bw() +
  boxplot_theme #+

p
  
readr::write_rds(p, file = './data/combined_figures/ammonia_island_shore_boxplots.rds')

ggsave(plot = p, 
       filename = './output/plots/ammonia_shore_bxplt_21.22.23.png',
       dpi = 1000,
       width = 7,
       height = 5)

```

## Model and plot nitrate + nitrite

```{r Nitrite_plus_Nitrate, warning=F, message=F, eval=T}
#| code-fold: true

formula <-
  Nitrite_plus_Nitrate ~
  Year * 
  Island_shore

sp.mod <- splm(formula = formula, 
               data = nutrients,
               random = ~ Site,
               spcov_type = 'gaussian') 

summary(sp.mod)

em.nn <- emmeans::emmeans(sp.mod, ~ Year)

print(em.nn)

pairs(em.nn)

plot(pairs(em.nn)) +
  theme_bw() +
  geom_vline(xintercept = 0)

em.nn <- emmeans::emmeans(sp.mod, ~ Island_shore, by = "Year")

print(em.nn)

pairs(em.nn)

plot(pairs(em.nn)) +
  theme_bw() +
  geom_vline(xintercept = 0)

## 1) Tidy pairwise tests for ggsignif
pairs_df <- 
  pairs(em.nn) %>%
  summary(adjust = "tukey") %>%
  as_tibble() %>%
  separate(contrast, into = c("group1", "group2"), sep = " - ") %>%
  mutate(
    # make sure group names match x-axis categories exactly
    group1 = as.character(group1),
    group2 = as.character(group2),
    label  = paste0("p=", scales::pvalue(p.value))) %>% 
  mutate(label = ifelse(label == 'p=<0.001', 'p<0.001', label))

## 2) Give each comparison a y position per Year so brackets don’t overlap
yr_tops <- nutrients %>%
  group_by(Year) %>%
  summarise(y_top = max(Nitrite_plus_Nitrate, na.rm = TRUE), .groups = "drop")

range_y <- diff(range(nutrients$Nitrite_plus_Nitrate, na.rm = TRUE))

pairs_df <- pairs_df %>%
  left_join(yr_tops, by = "Year") %>%
  group_by(Year) %>%
  mutate(
    # stack 3 comparisons above the tallest box for that Year
    y_position = y_top + (row_number()) * (0.06 * range_y)
  ) %>%
  ungroup()

## 3) Plot with ggsignif (manual = TRUE)
p <-
  ggplot(nutrients, aes(x = Island_shore, y = Nitrite_plus_Nitrate, color = Island_shore)) +
  geom_boxplot(outlier.shape = NA, alpha = 0.9, size=0.8) +
  geom_quasirandom(size = 2.5, alpha = 0.3, stroke = NA) +
  scale_color_manual(values = shore_colors) +
  facet_wrap(~ Year) +
  ggsignif::geom_signif(
    data = pairs_df,
    aes(xmin = group1, xmax = group2, annotations = label, y_position = y_position),
    inherit.aes = FALSE,
    manual = TRUE,          # <-- REQUIRED when mapping via data frame
    tip_length = 0.01,
    vjust = 0.2,
    textsize = 3.6) +
  labs(y = expression("Nitrite+Nitrate ("*mu*"M)"),
       x = 'Island Shore') +
  theme_bw() +
  boxplot_theme #+

p
  
readr::write_rds(p, file = './data/combined_figures/nitrate_nitrite_island_shore_boxplots.rds')

ggsave(plot = p, 
       filename = './output/plots/nitrate_nitrite_shore_bxplt_21.22.23.png',
       dpi = 1000,
       width = 7,
       height = 5)

```

## Model and plot phosphate

```{r phosphate, warning=F, message=F, eval=T}
#| code-fold: true

formula <-
  Phosphate ~
  Year * 
  Island_shore

sp.mod <- splm(formula = formula, 
               data = nutrients,
               random = ~ Site,
               spcov_type = 'gaussian') 

summary(sp.mod)

em.phosphate <- emmeans::emmeans(sp.mod, ~ Year)

print(em.phosphate)

pairs(em.phosphate)

plot(pairs(em.phosphate)) +
  theme_bw() +
  geom_vline(xintercept = 0)

em.phosphate <- emmeans::emmeans(sp.mod, ~ Island_shore, by = "Year")

print(em.phosphate)

pairs(em.phosphate)

plot(pairs(em.phosphate)) +
  theme_bw() +
  geom_vline(xintercept = 0)

## 1) Tidy pairwise tests for ggsignif
pairs_df <- 
  pairs(em.phosphate) %>%
  summary(adjust = "tukey") %>%
  as_tibble() %>%
  separate(contrast, into = c("group1", "group2"), sep = " - ") %>%
  mutate(
    # make sure group names match x-axis categories exactly
    group1 = as.character(group1),
    group2 = as.character(group2),
    label  = paste0("p=", scales::pvalue(p.value))) %>% 
  mutate(label = ifelse(label == 'p=<0.001', 'p<0.001', label))

## 2) Give each comparison a y position per Year so brackets don’t overlap
yr_tops <- nutrients %>%
  group_by(Year) %>%
  summarise(y_top = max(Phosphate, na.rm = TRUE), .groups = "drop")

range_y <- diff(range(nutrients$Phosphate, na.rm = TRUE))

pairs_df <- pairs_df %>%
  left_join(yr_tops, by = "Year") %>%
  group_by(Year) %>%
  mutate(
    # stack 3 comparisons above the tallest box for that Year
    y_position = y_top + (row_number()) * (0.06 * range_y)
  ) %>%
  ungroup()

## 3) Plot with ggsignif (manual = TRUE)
p <-
  ggplot(nutrients, aes(x = Island_shore, y = Phosphate, color = Island_shore)) +
  geom_boxplot(outlier.shape = NA, alpha = 0.9, size=0.8) +
  geom_quasirandom(size = 2.5, alpha = 0.3, stroke = NA) +
  scale_color_manual(values = shore_colors) +
  facet_wrap(~ Year) +
  ggsignif::geom_signif(
    data = pairs_df,
    aes(xmin = group1, xmax = group2, annotations = label, y_position = y_position),
    inherit.aes = FALSE,
    manual = TRUE,          # <-- REQUIRED when mapping via data frame
    tip_length = 0.01,
    vjust = 0.2,
    textsize = 3.6) +
  labs(y = expression("Phosphate ("*mu*"M)"),
       x = 'Island Shore') +
  theme_bw() +
  boxplot_theme #+

p
  
readr::write_rds(p, file = './data/combined_figures/phosphate_island_shore_boxplots.rds')

ggsave(plot = p, 
       filename = './output/plots/phosphate_shore_bxplt_21.22.23.png',
       dpi = 1000,
       width = 7,
       height = 5)
```
